+++ Entidades/Productos.cs
@@ -13,8 +13,8 @@ namespace ProyectoParcial.Entidades
         public int ProductoId { get; set; }
         public string Descripcion { get; set; }
         public int Existencia { get; set; }
-        public int Costo { get; set; }
-        public int ValorInventario { get; set; }
+        public decimal Costo { get; set; }
+        public decimal ValorInventario { get; set; }
 
         public Productos()
         {
